requirements.txt
kopf~=1.35.3
kubernetes~=20.13.0
python-dateutil>=2.8.2
